+++ Assets/Scripts/Map/Units/BaseUnit.cs
@@ -26,7 +26,7 @@ namespace Map.Unit
         const int _buildRange = 2;
         const bool _buildable = false;
         const int _baseCost = int.MaxValue;
-        const float _baseManaMoveRatio = float.MaxValue;
+        const int _baseManaMoveRatio = int.MaxValue;
         const int _baseManaAttackCost = int.MaxValue;
         readonly List<IEffect> _activeEffects = new List<IEffect>();
         IPlayer _owner;
@@ -60,7 +60,7 @@ namespace Map.Unit
         public int BuildRange { get { return _buildRange; } }
         public bool Buildable { get { return _buildable; } }
         public int Cost { get { return _baseCost; } }
-        public float ManaMoveRatio { get { return _baseManaMoveRatio; } }
+        public int ManaMoveRatio { get { return _baseManaMoveRatio; } }
         public int ManaAttackCost { get { return _baseManaAttackCost; } }
         public List<IEffect> ActiveEffects { get { return _activeEffects; } }
         public Sprite Icon { get { return icon; } }
